Fix FICO score precision issue: Convert Customer.ficoScore from Integer to BigDecimal
- Update Customer entity to use BigDecimal for ficoScore field to maintain COBOL COMP-3 precision
- Update AccountDto, AccountViewResponse, and CustomerDto to support BigDecimal ficoScore
- Ensure precise financial calculations match COBOL packed decimal behavior
- Critical fix for ReportMenuServiceTest validation and COBOL-to-Java migration compliance

+++ backend/src/main/java/com/carddemo/dto/AccountDto.java
@@ -235,9 +235,10 @@ public class AccountDto {
     /**
      * Customer FICO credit score.
      * Maps to CUST-FICO-CREDIT-SCORE from CVCUS01Y.cpy (PIC 9(03)).
+     * Uses BigDecimal for precise COBOL numeric handling.
      */
     @JsonProperty("customerFicoScore")
-    private Integer customerFicoScore;
+    private BigDecimal customerFicoScore;
 
     // ============ CALCULATED FIELDS ============
